coordinator: move peerdiscovery from enterprise
